Move WhyCommand to namespace matching folder structure
